+++ test/frame.js
@@ -18,7 +18,7 @@ describe('Navit.frame.*', () => {
       .listen(17345, err => {
         if (err) return done(err);
         // Init phantom before execute first test
-        browser.run(done);
+        browser.then(done);
       });
   });