+++ test/get.js
@@ -21,7 +21,7 @@ describe('Navit.get.*', () => {
       .listen(17345, err => {
         if (err) return done(err);
         // Init phantom before execute first test
-        browser.run(done);
+        browser.then(done);
       });
   });