+++ test/tab.js
@@ -19,7 +19,7 @@ describe('Navit.tab.*', () => {
       .listen(17345, err => {
         if (err) return done(err);
         // Init phantom before execute first test
-        browser.run(done);
+        browser.then(done);
       });
   });