+++ test/test.js
@@ -19,7 +19,7 @@ describe('Navit.test.*', () => {
       .listen(17345, err => {
         if (err) return done(err);
         // Init phantom before execute first test
-        browser.run(done);
+        browser.then(done);
       });
   });
 
@@ -319,8 +319,8 @@ describe('Navit.test.*', () => {
     });
   });
 
-  after(function (done) {
+  after(async () => {
     server.close();
-    browser.exit(done);
+    await browser.exit();
   });
 });